make some changes
