+++ .gitignore
@@ -12,8 +12,8 @@
 # Generated files
 .docusaurus
 .cache-loader
-/static/js/contributors.json
-/static/js/doc-contributors.json
+/static/json/contributors.json
+/static/json/doc-contributors.json
 
 # Misc
 .DS_Store
