+++ src/pages/team/index.js
@@ -1,4 +1,4 @@
-import React, { useState, useEffect }  from 'react';
+import React, { useState }  from 'react';
 import useIsBrowser from '@docusaurus/useIsBrowser';
 import useBaseUrl from '@docusaurus/useBaseUrl';
 import config from "../team/config.json";
@@ -14,14 +14,43 @@ export default function() {
     const language = isBrowser && location.pathname.indexOf('/zh-CN/') === 0 ? 'zh-CN' : 'en';
     const dataSource = config?.[language];
 
-    useEffect(() => {
-        if (isBrowser) {
-            const script = document.createElement('script');
-            script.type = 'module';
-            script.src = '../js/contributors.js';
-            document.body.appendChild(script);
+    const contributors = isBrowser && readTextFile("../json/contributors.json", function (text) {
+        let data = JSON.parse(text);
+        let eles = document.getElementById("conID");
+        for (let i = 0; i < data.length; i++) {
+            let lid = document.createElement("li");
+            lid.className = "mb-2";
+            lid.innerHTML = "<a href='"+data[i].html_url+"' target='_blank'>" +
+                "<img src='"+data[i].avatar_url+"'/>" + "<span>"+data[i].login+"</span>" +
+                "</a>";
+            eles.appendChild(lid);
         }
-    }, [isBrowser])
+    });
+
+    let doc_contributors = isBrowser && readTextFile("../json/doc-contributors.json", function (text) {
+        let data = JSON.parse(text);
+        let eles = document.getElementById("docID");
+        for (let i = 0; i < data.length; i++) {
+            let lid = document.createElement("li");
+            lid.className = "mb-2";
+            lid.innerHTML = "<a href='"+data[i].html_url+"' target='_blank'>" +
+                "<img src='"+data[i].avatar_url+"'/>" + "<span>"+data[i].login+"</span>" +
+                "</a>";
+            eles.appendChild(lid);
+        }
+    });
+
+    function readTextFile(file, callback) {
+        let rawFile = new XMLHttpRequest();
+        rawFile.overrideMimeType("application/json");
+        rawFile.open("GET", file, true);
+        rawFile.onreadystatechange = function () {
+            if (rawFile.readyState === 4 && rawFile.status == "200") {
+                callback(rawFile.responseText);
+            }
+        }
+        rawFile.send(null);
+    }
 
     return (
         <Layout>
